Don't try installing into /etc, it doesn't work with virtualenv.
When attempting to run 'python setup.py install' using a virtualenv, it
previously failed with:

  error: could not create '/etc/git-buildpackage': Permission denied

This commit removes that file from the list of files to be installed so that it
is possible to install this package in a virtualenv.

It is noted that upstream (https://git.sigxcpu.org/cgit/git-buildpackage) does
not attempt to install anything to /etc.

@todo Are any corresponding documentation updates required?  Should the file
instead be copied into a documentation directory?