pq-rpm: Allow to preserve the patch name on import/export.
This is based on upstream commit ID 8241459f which adds the support to 'pq'.

To apply the change, it was also necessary to add support to 'pq-rpm' for
"Gbp-Pq:" command tags, since it only supported "Gbp" ones previously, but
apply_and_commit_patch() generates "Gbp-Pq: Name" ones.

It was also necessary to add in filtering out of those lines from the generated
patch files.  This was previously omitted for "Topic", although that doesn't
actually seem to be used anyway.

@todo Should update documentation and tests as per the change to 'pq'.