Don't mangle commit comment lines starting with "From ".
Previously, when exporting .patch files, a ">" would be prepended to any line
starting with "From " (including the space) in the commit comment.

This fix is based on the way that the email.message.as_string() documentation
recommends avoiding a this specific form of mangling.

Once the mangling has occurred, it appears that importing the patches again will
result in lines starting with ">" being discarded.

@todo Should add/update a unit test to cover this.