RPM: Define _in_git_buildpackage when parsing .spec file.
In some cases, it is useful to be able to have the RPM .spec file behave
differently when it is being parsed by git-buildpackage as opposed to when an
actual build is being performed.

For example, if the .spec file uses Lua to read a source file and generate part
of the .spec file content from it, but the source file can only be found when
_topdir is defined correctly, and the generated part of the .spec file isn't
relevant to git-buildpackage's operation, it is simpler to just skip the Lua
code when running under git-buildpackage.

This commit modifies _parse_filtered_spec() to write "%define
_in_git_buildpackage 1" to the top of the temporary, filtered .spec file.  A
.spec file may then use the following construct to hide part of the .spec file
from git-buildpackage:

  %if ! 0%{?_in_git_buildpackage}
  ...
  %endif

@todo Need to document this.